main/python/requirements.txt
absl-py==2.1.0
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
immutabledict==4.2.0
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe>=2.1.0
mysql-connector-python==9.0.0
nltk>=3.8.1
numpy==1.24.4
ortools==9.11.4210
pandas==2.0.3
protobuf==5.26.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
regex==2024.9.11
scikit-learn>=1.3.2
scipy>=1.9.1
six==1.16.0
threadpoolctl==3.5.0
tqdm==4.66.5
tzdata==2024.2
Werkzeug==3.0.4
zipp==3.20.2
gunicorn==23.0.0
waitress==3.0.1
sqlalchemy==2.0.36
h5py==3.12.1
